20200808  物损定损功能搞定！！

+++ cBSPlus/src/main/java/com/cninsure/cp/entity/PublicOrderEntity.java
@@ -49,7 +49,7 @@ public class PublicOrderEntity implements Serializable {
 	/**报案号**/
 	public String baoanNo;
 	/**业务品种id**/
-    public long bussTypeId;
+    public Integer bussTypeId;
 	/**业务品种名称**/
     public String bussTypeName;
